+++ Common/Accessibility/ItemAutoReuse.cs
@@ -4,30 +4,19 @@ using TerrariaOverhaul.Core.Configuration;
 
 namespace TerrariaOverhaul.Common.Accessibility;
 
-public class ItemAutoReuse : GlobalItem
+// Before Terraria 1.4.4 added global auto-reuse as an opt-in for players,
+// Overhaul used to modify Item.autoReuse in SetDefaults() and add a small speed penalty to items it forced auto-reuse for:
+//	if (ForceAutoReuse && (!item.autoReuse && !item.channel) {
+//		item.useTime += 2;
+//		item.useAnimation += 2;
+//	}
+// Nowadays we just override 'Main.SettingsEnabled_AutoReuseAllItems' so that the option is enabled by default.
+public sealed class AutoReuseOverrideSystem : ModSystem
 {
 	public static readonly ConfigEntry<bool> ForceAutoReuse = new(ConfigSide.Both, "Accessibility", nameof(ForceAutoReuse), () => true);
 
-	public override void SetDefaults(Item item)
+	public override void PreUpdatePlayers()
 	{
-		if (!ForceAutoReuse) {
-			return;
-		}
-
-		if (item.autoReuse || item.channel) {
-			return;
-		}
-
-		//TODO: Implement through APIs instead.
-		if (item.ModItem?.Mod is { Name: "ClickerClass" }) {
-			return;
-		}
-
-		item.autoReuse = true;
-
-		// Prior to Terraria 1.4.4 enabling global auto-reuse as an opt-in for players,
-		// Overhaul used to add a small speed penalty to items it forced auto-reuse for:
-		//item.useTime += 2;
-		//item.useAnimation += 2;
+		Main.SettingsEnabled_AutoReuseAllItems |= ForceAutoReuse;
 	}
 }